fix(v4l2): ext-control value sits at packed offset 12, not 16
v4l2_ext_control's value union is __attribute__((packed)) — on the
64-bit UABI it lands at offset 12 and the struct totals 20 bytes. The
Go mirror declared it as a naturally-aligned int64, moving the value to
offset 16: the driver read the alignment padding (zero) instead.

Every int-valued codec control was therefore set to 0: BITRATE clamped
to the driver minimum (0.12 Mbps streams at a 2 Mbps config — the
90°/270° rotation path looked terrible), I_PERIOD/GOP_SIZE no-op (the
'software IDR pacing' workaround existed for this), REPEAT_SEQ_HEADER
never engaged (the SPS-cache injection existed for this). FORCE_KEY_FRAME
is a button control that ignores the value, which is why ext-ctrls
seemed to 'work' and masked the breakage. Layout pinned by compiler-
verified offsets from linux/videodev2.h (value @12, size 20, controls
ptr @24 in v4l2_ext_controls).

+++ internal/v4l2/encode.go
@@ -136,7 +136,7 @@ func OpenM2MEncoder(path string, width, height uint32, opts M2MEncoderOptions) (
 	// buffer-less context. I_PERIOD/GOP keeps keyframes periodic —
 	// without it mid-stream RTSP joins and recording segmentation break.
 	setCtrl := func(id uint32, value int32) {
-		_ = setExtCtrl(file.Fd(), id, int64(value))
+		_ = setExtCtrl(file.Fd(), id, value)
 	}
 	if opts.Bitrate > 0 {
 		setCtrl(cidVideoBitrate, opts.Bitrate)
